Update config.js
Edit to line 53 in the comment. Change example url from "https://myOrg.maps.arcgis.com/" to "https://myOrg.maps.arcgis.com" because if using OAuth for authentication, line 164 in mapBookConfigLoader.js will be equal to false (because the PortalURL does not match the same format as idObject.credentials[i].server) when it should read true.